EAR 4.2 initial release

+++ src/common/system/file.h
@@ -10,9 +10,9 @@
 * BSC Contact   mailto:ear-support@bsc.es
 * Lenovo contact  mailto:hpchelp@lenovo.com
 *
-* This file is licensed under both the BSD-3 license for individual/non-commercial
-* use and EPL-1.0 license for commercial use. Full text of both licenses can be
-* found in COPYING.BSD and COPYING.EPL files.
+* EAR is an open source software, and it is licensed under both the BSD-3 license
+* and EPL-1.0 license. Full text of both licenses can be found in COPYING.BSD
+* and COPYING.EPL files.
 */
 
 #ifndef _FILE_H
@@ -45,7 +45,7 @@
 int file_lock(int fd);
 
 /* Tries to acquire the lock with file_lock with a max a attemps = tiemoput, expressed as loop iterations, not time . Returns true when the lock has been acquired*/
-int file_lock_timeout(int fd, uint timeout);
+int file_lock_timeout(int fd, ulong timeout);
 
 /** Creates a file to be used  as lock. It doesn't locks the file */
 int file_lock_create(char *lock_file_name);
